add mark splitting for string

+++ src/natom.rs
@@ -25,7 +25,7 @@ impl NAtom {
         }
     }
 
-    pub(crate) fn size(&self) -> usize {
+    pub(crate) fn size(&self) -> u32 {
         1
     }
